pyproject.toml
[project]
name = "terminusgps-notifier"
description = "Terminus GPS Notification Dispatch Microservice"
readme = "README.md"
version = "3.2.0"
authors = [
    { name = "Blake Nall", email = "blake@terminusgps.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
]
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.5.0",
    "django>=6.0.2",
    "python-terminusgps>=51.0.0",
    "redis[hiredis]>=6.4.0",
    "terminusgps-payments>=8.3.0",
    "uvicorn>=0.35.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary,pool]>=3.2.11",
    "types-aioboto3[pinpoint-sms-voice-v2]>=15.5.0",
]

[project.urls]
Documentation = "https://docs.terminusgps.com"
Repository = "https://github.com/terminusgps/terminusgps-notifier"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["terminusgps_notifier/"]

[tool.pyright]
strictListInference = false
strictDictionaryInference = false
strictSetInference = false
typeCheckingMode = "basic"

[tool.ruff]
exclude = [
    ".git",
    ".gitignore",
    ".venv",
    "venv",
    ".vscode",
    "__pycache__",
    "node_modules",
    "site-packages",
    "dist",
    "build",
    "db.sqlite3",
]
line-length = 79
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = [
    "E4",
    "E7",
    "E9",
    "F",
    "DJ",
]
ignore = ["DJ001", "F821"]
fixable = [ "ALL" ]
unfixable = []
extend-per-file-ignores = {"*/migrations/*.py" = ["E501"]}

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = false
skip-magic-trailing-comma = true

[dependency-groups]
dev = [
    "sphinx-rtd-theme>=3.0.2",
    "sphinx>=8.2.3",
]
